refactor: align UI padding and rename campsite provider file
- Updated paddings and layout spacing in various widgets to ensure consistent alignment.
- Renamed campsite_provider.dart → campsite_providers.dart for better clarity and convention.
- Minor cleanup for visual consistency across compact and wide layouts.
