+++ lib/features/campsite/presentation/widgets/result_summary_banner.dart
@@ -27,28 +27,26 @@ class ResultSummaryBanner extends StatelessWidget {
             ? '$filteredCount of $totalCount campsites match your filters'
             : '$totalCount campsites available';
 
-    return Padding(
-      padding: const EdgeInsets.symmetric(vertical: AppSizes.paddingSM),
-      child: Card(
-        color: Theme.of(context).colorScheme.primary.alphaF(0.2),
-        shape: RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(AppSizes.radius),
-        ),
-        elevation: 0,
-        child: Padding(
-          padding: const EdgeInsets.all(AppSizes.paddingSM),
-          child: Row(
-            mainAxisSize: MainAxisSize.min,
-            children: [
-              Icon(
-                isFiltered ? Icons.filter_alt_outlined : Icons.info_outline,
-                size: AppSizes.icon,
-                color: Theme.of(context).colorScheme.onSurface,
-              ),
-              const SizedBox(width: AppSizes.spaceSM),
-              Text(label, style: textTheme.bodyMedium),
-            ],
-          ),
+    return Card(
+      margin: const EdgeInsets.symmetric(vertical: AppSizes.paddingSM),
+      color: Theme.of(context).colorScheme.primary.alphaF(0.2),
+      shape: RoundedRectangleBorder(
+        borderRadius: BorderRadius.circular(AppSizes.radius),
+      ),
+      elevation: 0,
+      child: Padding(
+        padding: const EdgeInsets.all(AppSizes.paddingSM),
+        child: Row(
+          mainAxisSize: MainAxisSize.min,
+          children: [
+            Icon(
+              isFiltered ? Icons.filter_alt_outlined : Icons.info_outline,
+              size: AppSizes.icon,
+              color: Theme.of(context).colorScheme.onSurface,
+            ),
+            const SizedBox(width: AppSizes.spaceSM),
+            Text(label, style: textTheme.bodyMedium),
+          ],
         ),
       ),
     );
