+++ lib/features/favourites/presentation/screens/favourite_campsites_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:wander_nest/features/campsite/domain/entities/campsite.dart';
-import 'package:wander_nest/features/campsite/presentation/providers/campsite_provider.dart';
+import 'package:wander_nest/features/campsite/presentation/providers/campsite_providers.dart';
 import 'package:wander_nest/features/campsite/presentation/widgets/campsite_card.dart';
 import 'package:wander_nest/features/favourites/presentation/providers/favourites_provider.dart';
 import 'package:wander_nest/shared/widgets/animated_lists.dart';
